Were created the activities CharactersActitivy, and DetailActivity with de layouts activity_characters, activity_detail, and character_cards_layout.
It was added the Internet permission in the AndroidManifest.

Were added components in the layouts activity_main, activity_characters, activity_detail, and character_cards_layout.

It was added the drawable super_hero_gradient.

It was created the file dimems.xml.

Were added components in the files dimens, and strings.

Were added libraries in the build.gradle.